+++ cloudslang-compiler/src/main/java/io/cloudslang/lang/compiler/modeller/transformers/AbstractOutputsTransformer.java
@@ -87,29 +87,7 @@ public abstract class AbstractOutputsTransformer  extends InOutTransformer {
         outputs.add(element);
     }
 
-    Output createPropOutput(Map.Entry<String, Map<String, Serializable>> entry) {
-        Map<String, Serializable> props = entry.getValue();
-        validateKeys(entry, props);
-        // default is sensitive=false
-        String outputName = entry.getKey();
-        boolean sensitive = props.containsKey(SENSITIVE_KEY) && (boolean) props.get(SENSITIVE_KEY);
-        Serializable value = props.get(VALUE_KEY);
-        if (value == null) {
-            return createRefOutput(outputName, sensitive);
-        }
-
-        return createOutput(outputName, value, sensitive);
-    }
-
-    private void validateKeys(Map.Entry<String, Map<String, Serializable>> entry, Map<String, Serializable> props) {
-        for (String key : props.keySet()) {
-            if (!KNOWN_KEYS.contains(key)) {
-                throw new RuntimeException("Key: " + key + " in output: " + entry.getKey() + " is not a known property");
-            }
-        }
-    }
-
-    private Output createOutput(String outputName, Serializable outputExpression, boolean sensitive){
+    Output createOutput(String outputName, Serializable outputExpression, boolean sensitive){
         Accumulator accumulator = extractFunctionData(outputExpression);
         return new Output(
                 outputName,
@@ -119,7 +97,7 @@ public abstract class AbstractOutputsTransformer  extends InOutTransformer {
         );
     }
 
-    private Output createRefOutput(String rawOutput, boolean sensitive) {
+    Output createRefOutput(String rawOutput, boolean sensitive) {
         return new Output(rawOutput, ValueFactory.create(transformNameToExpression(rawOutput), sensitive));
     }
